switched to bitmasks

+++ src/CreateGraph.h
@@ -19,11 +19,11 @@ using namespace cv;
 //TODO: Fix the weight from int to float
 struct edge{
   public:
-    int32_t Vertex;
-    int32_t Weight;
+    uint32_t Vertex;
+    uint64_t Weight;
 };
 
-int ImagetoGraphSerial(Mat image, edge *EdgeList, int32_t *VertexList, int32_t *BitEdgeList);
+int ImagetoGraphSerial(Mat image, edge *EdgeList, uint32_t *VertexList, uint64_t *BitEdgeList);
 __global__ void ImagetoGraph(cv::cuda::GpuMat Image, int32_t *VertexList, edge *EdgeList, int32_t *BitEdgeList, int32_t *FlagList, int32_t pitch, int32_t channels);
 
 #endif //FELZENSWALB_CREATEGRAPH_H
